Fixes DEVELOPER-934 (adding social links)
I've added the ability to add different accounts. That seems to be the
most flexible way of doing this. The user will click the button to add
an account which will then add a new group of account fields, similar to
how the spec and version groups work. The user will need to add the URL
for the account homepage and also the account name (i.e.
https://twitter.com and wildflyas). When the DOAP file is generated the
new account(s) will be at the bottom of the DOAP file.